requirements.txt
pillow
numpy
scipy
matplotlib
pyqt5
pyqt5-stubs
